Move @cpu/@sys expansion back to Lookup
We cannot expand in Readlink because we already reported the length of
the link on an earlier stat without taking expansion in to account. But
to report the correct length we would have to fetch data for symlinks
and at the time we fetch the symlink attributes fetch we don't yet know
if the object is going to be a symlink.

Revert parts of e53b3cb7dc0e563a1ce74439267cdd4e1c26b8b6