Fix a server crash during resolution.
During resolution we hit a case where a site that supposedly
participated in a create got the create again. I assume the order
of operations here was some variation of,

        CREATE a/foo
        RENAME a/foo b/foo
        *conflict*

Then repair 'a' heuristics find 'foo' is missing and attempt to
re-create.  Mark parent inconflict, this should trigger resolution
into checking the destination of the rename first or else mark the
parent directory as requiring manual repair.

"pathname not leftmost" message was hiding real error messages in
the repair tool.