Avoid using server IP address in storeids.
During resolution a server would initialise the ViceStoreId.Host field to
the IP address of the server. Clients used to do this, but are using a
random 32-bit value chosen during rvm initialization.

We only compare the storeids for equality/inequality so we can replace it
with any other (fairly constant and probably unique) value, so I'm using
serverid.

This change actually renames Host to HostId, because I wanted to be sure
we never compared against a server IP address.