Use non-local 'address' for volume lock
When reintegrating, do not use the local ip address of the server for
the volume lock because different parts of the resolution process use
this to check if the lock was taken by the resolution process.

We can't use '0' either because that is treated as the equivalent of
no lock, so we'll use -1 (255.255.255.255) because that will never be
the local ipv4 of a Coda server.