Allow automated license categorization.
Although Coda consists of both of GPLv2 and LGPLv2 licensed code, the
governing license for the body of code is GPLv2, while seperable
libraries such as LWP, RPC2 and RVM overall fall under the LGPLv2
license. But automated systems are typically unable to correctly parse
out the finer details and classify Coda as unlicensed.

- moved the higher level licensing related comments to the COPYING file
- made LICENSE a verbatim copy of GNU General Public License, version 2
  (http://www.gnu.org/licenses/gpl-2.0.txt)
- added LICENSE.lgpl, the GNU Library General Public License, version 2.0
  (http://www.gnu.org/licenses/lgpl-2.0.txt)