Fixed a buffer overflow in 'hoard'
When 'hoard' is called with a long an erroneous command line then a
buffer overflow occurs. This commit fixes this kind of overflow.

The reason is because the function 'error' (which is called by
'parse_error') tries to copy all the remaining characters of the
commandline (following the parse error) into a buffer with a size
of 240 bytes.

This commit replaces 'vsprintf' with a 'vsnprintf' so in case the
command line is too long it gets truncated for the error message.