Rephrased NOP in 'vice-setup-user'
My version of bash (which is 4.3.46(1)) doesn't like a single
semicolon on a line of its own. So when executing 'vice-setup-user'
it says:

        syntax error near unexpected token `;'

Since there only has to be a "nop-command" I replaced the
semicolon with an "echo to dev null". This can be handled by
bash and has the same effect.