Revert "Drop need for PrimaryHost parameter from ViceFetch"
This reverts commit ad428ac575113b75e63ff6d10cf00fdc0003025e.

I was assuming that when the client sets OMITSE in a multirpc request we
didn't send parameters and as such could tell on the server side that
there is no side-effect intialized. However, the parameters are sent on
the initial packet sent in the (longer lived) connection and as such
don't give a hint if a sideeffect is active for the current request.