Fix regression where we failed to connect to update server.
The removal of multiple Coda server (on one machine) support introduced a bug
where we were using Host to indicate both the Update server host as well as the
local Coda server. As a result we failed to connect if the Update server is not
on the same machine as the Coda server we are notifing.