Add commandline option to enable 9pfs support
Default to not enabled because the mariner socket is world accessible
for codacon, but when we talk to it using a 9pfs client we can now
assume the identity and access of any locally authenticated Coda user.

(i.e. the user has to have used clog on the local /coda mount because we
don't support authentication through the 9pfs emulation layer yet)

We may be able to fix this once authentication through the 9pfs mount is
supported, otherwise we may have to add a 'root' check for the mariner
port and only allow 9pfs clients that connect through a unix domain
socket from a system administrator account.