Changes to SFTP to avoid sending too much data
On the server, don't send gratuitous acknowledgments when we receive
duplicate data. Even if our acknowledgments were lost we will resend
when we are not able to advance the send window.
On the client, send only a single worried packet when the window is
closed to reduce how many data packets may be queued on the network.