AC_FUNC_MMAP configure check is too strict
This macro checks if the mmap function exists and works. However this
fails when we try to build Coda on a filesystem (userspace or network)
that doesn't support mmap. We can always try to implement a runtime
workaround if with malloc and read although it would be hard to allocate
to a predictable and consistent address.