Modify pioctl support to use virtual files in Coda's root
This should allow us to use ioctls on systems which needed hacks to
support ioctls, such as Windows. Or when there is no ioctl equivalent
when Venus exposes the file system as a Plan9 file server. It also
allows us to remove some ugly bits of the Coda kernel module.

This doesn't solve the 32-bit/64-bit issues (32-bit Venus and 64-bit
cfs/clog or vice versa) because we still are passing opaque binary
structures back and forth.