Fix buffer overflow in rvmutl parser.
We were copying an rvm_offset_t to a double assuming that both were the
same length. But on a 64-bit system the rvm_offset_t ends up being 128
bits and so we overflow the size of the buffer.

Should rvm_offset_t be reduced to a 64-bit value instead?