Avoid 100% CPU use after a mariner socket closes.
I tried a couple of approaches and this seemed the most reliable at
picking up when the peer has closed the connection, in case of named
pipes it doesn't seem like efd is set, and we cannot just exit whenever
we get a 0-byte read. So the middle ground seems to be a 0-byte read
after select returns that the socket is readable.