Fix for uv_udp_send regression in libuv 1.27.0
Older versions of libuv didn't really support sending to unix domain
sockets, we had to add a dummy peer to avoid a null-ptr dereference.

However in libuv-1.27.0 support for connected UDP sockets was added and
now it tries (and fails) to bind to the dummy peer address. To avoid
this bind attempt we have to pass a NULL ptr, which doesn't work with
older versions of libuv.