Read VRList instead of VRDB file when reloading the vrdb.
The VRDB file is a binary format that changes between 32-bit and 64-bit
systems. It is more reliable to simply parse the VRList file.

Right now we still make a new binary VRDB file in createvol_rep before
we rename VRList.tmp to VRList. This is to make sure the new file can be
read without errors and so that there still is a VRDB file for the older
servers that do not read the VRList file yet.