Fix race during gnutls_handshake
We were not calling peeloff_and_decrypt during the handshake, but when the
other side closes the connection after the handshake completes this could
arrive before we transition the connection state and we end up not getting
woken up for the EOF signal.

Call peeloff_and_decrypt after a successfull handshake to make sure we pick
up any packets that were received during, but not part of, the handshake.