During recovery, a Dying object may have Invalid type
A newly initialized object is (Runt,Invalid), then when we get the
attributes from the server both the state and type are updated. If we
shutdown or crash before getattr returns, we reclaim this runt object
during recovery by setting the state to Dying.

But if we shut down again before garbage collecting the fsobj, we find
the now (Dying,Invalid) object and crashed because it has no proper type.
However if an object is FsoDying it should not make any difference if the
VnodeType is Invalid because it will be garbage collected, so don't kill
the client when we see it.