Remove coda-src/vtools/up.cc
This has been in the source tree since the beginning of recorded Coda
history, but it has not been packaged in the distributed source tar
archives or compiled. There is no manpage and it contains 2 different
(conflicting) 'usage:' strings.

Its purpose seems to be to recursively copy a directory tree while
preserving mode and owner bits with the additional option of converting
directory ACLs from an old to a new format. We can use off-the-shelf
`cp` or `rsync` to do the recursive copy and I don't believe the old
ACL format exists anymore so the extra conversion step is not necessary.